% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.data.frame.r
\name{add_stats.data.frame}
\alias{add_stats.data.frame}
\title{tidystats method for adding a tidy data frame to a tidy results list}
\usage{
\method{add_stats}{data.frame}(output, results, identifier = NULL,
  statistics = NULL, type = NULL, confirmatory = NULL, notes = NULL)
}
\arguments{
\item{results}{a tidy stats list.}

\item{identifier}{a character string identifying the model. Automatically created if not provided.}

\item{type}{a character string indicating the type of test. One of "hypothesis", "manipulation check", "contrast", "descriptives", or "other", can be abbreviated.}

\item{confirmatory}{was the analysis confirmatory? TRUE, FALSE, or NULL.}

\item{notes}{a character string to add additional information.}

\item{data}{output of a statistical test in a tidy data frame format.}
}
\description{
\code{add_stats.data.frame} is a function to add a tidy data frame of results to a tidy stats list. tidystats does not support all possible statistical tests, so it may not be able to produce tidy output of a statistical model. The best solution for now is to tidy the output of a statistical method yourself, creating a tidy data frame, and then use \code{add_stats}, which will call this function, to add it to the tidy stats list.
}
