% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.default.r
\name{add_stats.default}
\alias{add_stats.default}
\title{add_stats default function}
\usage{
\method{add_stats}{default}(results, output, identifier = NULL,
  notes = NULL, class = NULL, args = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{output}{Output of a statistical test.}

\item{identifier}{A character string identifying the model. Automatically
created if not provided.}

\item{notes}{A character string to add additional information. Some
statistical tests produce notes information, which will be overwritten if
notes are provided.}

\item{class}{A character string to indicate which function was used to
produce the output. See 'Details' for a list of supported functions.}

\item{args}{An optional list of additional arguments. Can be used to specify
how model results should be summarized.}
}
\description{
\code{add_stats.default} is the default add_stats function, which takes the
output of a statistical test, organizes the output, and adds it to a tidystats
list.
}
