% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats_manual.r
\name{add_stats_manual}
\alias{add_stats_manual}
\title{Manually add stats to tidy results list}
\usage{
add_stats_manual(data, results, identifier = NULL, type = NULL,
  description = NULL)
}
\arguments{
\item{data}{output of a statistical test in a tidy data frame format.}

\item{results}{a tidy stats list.}

\item{identifier}{a character string identifying the model. Automatically created if not provided.}

\item{type}{a character string indicating the type of test. One of "hypothesis", "manipulation check", "contrast", or "other", can be abbreviated.}

\item{description}{a character string to add additional information.}
}
\description{
\code{add_stats_manual} is a function to add a tidy data frame of statistical results to a tidy stats list. tidystats does not support all possible statistical tests, so it may not be able to produce tidy output of a statistical model. The best solution for now is to tidy the output of a statistical method yourself, and then use this function to add it to the tidy stats list.
}
