% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addstats.R
\name{addstats}
\alias{addstats}
\title{Addstats generic function}
\usage{
addstats(results, model, identifier, type = "", description = NULL)
}
\description{
\code{addstats} is a generic function to add a line of tidy stats to a tidy stats data.frame.
}
\examples{
results <- new_stats_data_frame()
ttest <- t.test(1:10, y = c(7:20))
results <- addstats(results, ttest, "ttest_x_y", "Hypothesis", "A t-test of x and y")
}
