% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome_lm.r
\name{awesome_lm}
\alias{awesome_lm}
\title{Awesome Linear Model Fitting}
\usage{
awesome_lm(formula, data, subset, weights, na.action, method = "qr",
  model = TRUE, x = FALSE, y = FALSE, qr = TRUE, singular.ok = TRUE,
  contrasts = NULL, offset, level = 0.95, ...)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{na.action}{a function which indicates what should happen when the data contain NAs.}

\item{method}{the method to be used; for fitting, currently only method = "qr" is supported; method = "model.frame" returns the model frame (the same as with model = TRUE, see below).}

\item{model, x, y, qr}{logicals. If TRUE the corresponding components of the fit (the model frame, the model matrix, the response, the QR decomposition) are returned.}

\item{singular.ok}{logical. If FALSE (the default in S but not in R) a singular fit is an error.}

\item{contrasts}{an optional list. See the contrasts.arg of model.matrix.default.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if more than one are specified their sum is used. See model.offset.}

\item{level}{the confidence level required.}

\item{...}{additional arguments to be passed to the low level regression fitting functions (see below).}
}
\value{
\code{awesome_lm} returns an object of class "\code{awesome_lm}" and "\code{lm}" or for multiple responses of class c("\code{awesome_mlm}", "\code{mlm}", "\code{awesome_lm}", "\code{lm}").

The functions \code{summary} and \code{anova} are used to obtain and print a summary and analysis of variance table of the results.
}
\description{
\code{awesome_lm} is based on \code{lm}, but also includes standardized coefficients and confidence intervals.
}
\details{
See \code{?lm} for more information. Confidence intervals are created using \code{confint}.
}
\examples{

# Annette Dobson (1990) "An Introduction to Generalized Linear Models".
# Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)

# Example 1: Standard regresssion
lm1 <- lm(weight ~ group)
summary(lm1)

# Example 2: 90\% confidence intervals
lm2 <- lm(weight ~ group, level = .9)
summary(lm2)

}
