% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Tidy correlation}
\usage{
correlation(identifier, type, model, description = NULL)
}
\description{
Create tidy correlation output.
}
\examples{
x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
y <- c( 2.6,  3.1,  2.5,  5.0,  3.6,  4.0,  5.2,  2.8,  3.8)

model <- cor.test(x, y)
correlation("M1", "hypothesis test", model)
}
