% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_table.R
\name{correlation_table}
\alias{correlation_table}
\title{Create a correlation table}
\usage{
correlation_table(correlations, triangle = "lower", p_values = NULL,
  diagonal = NULL, labels = NULL, digits = 2)
}
\arguments{
\item{correlations}{A data frame or matrix containing the correlations.}

\item{triangle}{A character string indicating whether the lower, upper, or
both halves should be visible, must be one of 'lower', 'upper', or 'both'.
You can specify just the initial letter.}

\item{p_values}{A data frame or matrix containing the p-values.}

\item{diagonal}{A vector of values to be placed on the diagonal. Default is
a vector of 1s.}

\item{labels}{A vector of strings containing the labels for each variable.}

\item{digits}{An integer indicating the number of decimal places.}
}
\description{
Function to create a correlation table.
}
\examples{
# Add examples

}
