% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_data.r
\name{count_data}
\alias{count_data}
\title{Count the number of observations}
\usage{
count_data(data, ..., by = NULL, na.rm = FALSE, pct = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{...}{One or more unquoted (categorical) column names from the data
frame, separated by commas.}

\item{by}{An optional character vector of column names to group by.}

\item{na.rm}{A boolean specifying whether missing values (including NaN)
should be removed.}

\item{pct}{A boolean indicating whether to calculate percentages instead of
proportions. The default is \code{FALSE}.}
}
\description{
\code{\link[=count_data]{count_data()}} returns the number and proportion of observations for
categorical variables.
}
\details{
Use the \code{by} argument to group the data, or alternatively pipe
grouped data created with \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\examples{
count_data(quote_source, source)
count_data(quote_source, source, sex)
count_data(quote_source, source, sex, na.rm = TRUE)
count_data(quote_source, source, sex, na.rm = TRUE, pct = TRUE)

# Use the by argument to calculate proportions within a group
count_data(quote_source, sex, by = "source")

}
