% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\docType{data}
\name{cox}
\alias{cox}
\title{Data of Wissink, Hoogendoorn, Brohmer, Verschoor, Krijnen, & Zeelenberg}
\format{A data frame with 200 rows and 67 variables:}
\usage{
cox
}
\description{
Replication of “Terror management and adults’ attachment to their parents: The safe haven remains” by CR Cox, J Arndt, T Pyszczynski, J Greenberg, A Abdollahi, S Solomon (2008, Journal of Personality and Social Psychology).
}
\references{
\url{https://osf.io/5tbxf/}
}
\keyword{datasets}
