% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_descriptives.r
\name{D}
\alias{D}
\alias{M}
\alias{SD}
\title{Report descriptives}
\usage{
D(list, identifier, var, group = NULL, statistic = NULL)

M(list, identifier, var, group = NULL)

SD(list, identifier, var, group = NULL)
}
\arguments{
\item{identifier}{An identifier.}

\item{var}{A character string indicating which variable you want to report the descriptive of.}

\item{group}{(Optional) A character string indicating which group you want to report the descriptive of.}

\item{statistic}{A character string indicating which statistic you want to report.}

\item{results}{A tidystats list.}
}
\description{
\code{D} is a general function that returns a single descriptives from a tidystats list.
}
\details{
\code{D} retrieves the descriptive of a specific variable (and group).
}
\examples{
# Read in a list of results
descriptives <- read_stats(system.file("descriptives.csv", package = "tidystats"))

# Generic example:
D(descriptives, identifier = "D1", var = "avoidance", statistic = "M")

# Shorter examples:
# Mean
M(descriptives, "D1", "avoidance")

# SD
SD(descriptives, "D1", "avoidance")

}
