% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_list_to_df.R
\name{descriptives_list_to_df}
\alias{descriptives_list_to_df}
\title{Convert a tidystats descriptives list to a data frame}
\usage{
descriptives_list_to_df(descriptives)
}
\description{
\code{descriptives_list_to_df} converts a tidystats descriptives list to a data frame
}
\examples{
descriptives_list_to_df(descriptives)

}
