% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.r
\name{frequencies}
\alias{frequencies}
\title{Calculate frequency statistics}
\usage{
frequencies(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{...}{A variable number of grouping variables. If provided, descriptives will be calculated
for each level of the grouping variables}

\item{na.rm}{Boolean to indicate whether missing data should be removed. The default is TRUE.}
}
\description{
\code{frequencies} returns a set of common frequency statistics (e.g., n, percentage).
}
\examples{
frequencies(sleep, group)

}
