% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.default.R
\name{inspect.default}
\alias{inspect.default}
\title{Inspect (a) statistical model(s) output}
\usage{
\method{inspect}{default}(results, ...)
}
\arguments{
\item{results}{The output of a statistical test.}

\item{...}{Unused}
}
\description{
\code{inspect} is a function to inspect the output of a statistical model via
the Viewer pane.
}
\examples{
# Run a statistical test
model <- t.test(extra ~ group, data = sleep)

# Inspect the output
# inspect(model)

}
