% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.r
\name{list_to_df}
\alias{list_to_df}
\title{Convert a tidystats list to a data frame}
\usage{
list_to_df(results)
}
\description{
\code{list_to_df} converts a tidystats list to a data frame
}
\examples{
list_to_df(results)

}
