% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptives_data_frame.R
\name{new_descriptives_data_frame}
\alias{new_descriptives_data_frame}
\title{Create Empty Descriptives Data Frame}
\usage{
new_descriptives_data_frame()
}
\description{
This function allows you to create an empty data frame to store descriptives in.
}
\examples{
new_descriptives_data_frame()

}
