% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats_data_frame.R
\name{new_stats_data_frame}
\alias{new_stats_data_frame}
\title{Create Empty Stats Data Frame}
\usage{
new_stats_data_frame()
}
\description{
This function allows you to create an empty data frame to store the descriptive information and output of statistical models in.
}
\examples{
new_stats_data_frame()

}
