% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stats_list.R
\name{new_stats_list}
\alias{new_stats_list}
\title{Create Empty Stats Data Frame}
\usage{
new_stats_list()
}
\description{
This function allows you to create an empty data frame to store the output of statistical models in.
}
\examples{
new_stats_data_frame()

}
