% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename statistics columns}
\usage{
rename_columns(x)
}
\arguments{
\item{x}{The output of a statistical model, converted to a data frame.}
}
\description{
Renames the statistics columns.
}
\details{
This function matches each column name with an entry in a list of
statistical names and renames the column if necessary. The goal is to create
a naming scheme that makes statistical labels as consistent as possible.
}
