% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.data.frame.R
\name{report.data.frame}
\alias{report.data.frame}
\title{Report generic function}
\usage{
report.data.frame(results, identifier, term = NULL, statistic = NULL)
}
\description{
\code{report} is a generic function that reports a test in APA style.
}
