% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.htest.r
\name{report.htest}
\alias{report.htest}
\title{Report method for an htest object}
\usage{
\method{report}{htest}(results, identifier, term = NULL, statistic = NULL)
}
\description{
Reports an htest object in APA style.
}
\examples{
# Create empty results data frame
results <- new_stats_data_frame()

# Run a t-test
model <- t.test(1:10, y = c(7:20))

# Add model to results
add_stats(model, identifier = "M1", type = "hypothesis")

# Report model
report(results, "M1")
}
