% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report function}
\usage{
report(results, identifier, term = NULL, statistic = NULL)
}
\description{
\code{report} is a general function that reports a test in APA style by checking what kind of statistical test is supplied and calling its associated report function.
}
