% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_chi_squared.R
\name{report_chi_squared}
\alias{report_chi_squared}
\title{Report function for a chi-squared test}
\usage{
report_chi_squared(results, identifier, statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{statistic}{A character string identifying the exact statistic you want
to report.}
}
\description{
Function to report a chi-squared test in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "chi_square")
report(results, identifier = "chi_square", statistic = "p")

}
