% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_glm.R
\name{report_glm}
\alias{report_glm}
\title{Report method for generalized linear models}
\usage{
report_glm(results, identifier, group = NULL, term = NULL,
  term_nr = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{group}{A character string indicating the group containing the statistic
or statistics you want to report.}

\item{term}{A character string indicating the term you want to report.}

\item{term_nr}{A number indicating the term you want to report.}
}
\description{
Function to report a generalized linear models in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Set the default results list
options(tidystats_list = results)

# Report results
report("lm_simple", term = "conditionmortality salience")
report("lm_simple", term_nr = 2)
report("lm_simple", group = "model")

}
