% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p_value.R
\name{report_p_value}
\alias{report_p_value}
\title{Report p-value function}
\usage{
report_p_value(p_value)
}
\description{
A helper function to report p values in APA style.
}
\examples{
report_p_value(.45)
report_p_value(.002)
report_p_value(.00000122)
}
