% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_rma.R
\name{report_rma}
\alias{report_rma}
\title{Report method for metafor's rma models}
\usage{
report_rma(results, identifier, term = NULL, term_nr = NULL,
  statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{term}{A character string indicating which term you want to report the statistics of.}

\item{term_nr}{A number indicating which term you want to report the the statistics of.}

\item{statistic}{A character string of a statistic you want to extract from a model.}
}
\description{
Function to report a meta-analysis in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Set the default results list
options(tidystats_list = results)

# Example: regression term
# report("regression", term = "groupTrt")

}
