% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_statistic.r
\name{report_statistic}
\alias{report_statistic}
\title{Report a single statistic}
\usage{
report_statistic(statistic, value)
}
\arguments{
\item{statistic}{A character string indicating the kind of statistic (e.g.,
p, t, df).}

\item{value}{The value of the statistic to be reported.}
}
\description{
A helper function to report single statistics in APA style.
}
\examples{
report_statistic("p", 0.0001)
report_statistic("t", 4.2325)
report_statistic("r", 0.213)

}
