% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.r
\name{stat}
\alias{stat}
\title{Create a custom statistic}
\usage{
stat(
  name,
  value,
  symbol = NULL,
  subscript = NULL,
  interval = NULL,
  level = NULL,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{name}{A string specifying the name of the statistic.}

\item{value}{The numeric value of the statistic.}

\item{symbol}{A string specifying the symbol of the statistic to use when
reporting the statistic.}

\item{subscript}{A string specifying a subscript to use when reporting the
statistic.}

\item{interval}{A string specifying the type of interval if the statistic is
a ranged statistic (e.g., one with confidence intervals)/}

\item{level}{A numeric value between 0 and 1 indicating the level of the
interval.}

\item{lower}{The numeric value of the lower bound of the statistic.}

\item{upper}{The numeric value of the upper bound of the statistic.}
}
\description{
\code{stat()} can be used to create your own statistic to add to a custom
statistics object using \code{\link[=stats]{stats()}}.
}
