% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{stats}
\alias{stats}
\title{Create a custom statistics object}
\usage{
stats(method, statistics)
}
\arguments{
\item{method}{A string; specifying the path to type of method used to obtain
the statistics.}

\item{statistics}{A vector of statistics created with \code{stat()}.}
}
\description{
\code{stats} can be used to create your own statistics object to add to a
statistics list via \code{add_stats()}.
}
