% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_list_to_df.r
\name{stats_list_to_df}
\alias{stats_list_to_df}
\title{Convert a tidystats list to a data frame}
\usage{
stats_list_to_df(results)
}
\description{
\code{stats_list_to_df} converts a tidystats list to a data frame
}
\examples{
stats_list_to_df(results)

}
