% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.awesome_lm.r
\name{summary.awesome_lm}
\alias{summary.awesome_lm}
\title{Summarizing Awesome Linear Model Fits}
\usage{
\method{summary}{awesome_lm}(object, correlation = FALSE,
  symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{an object of class "lm", usually, a result of a call to lm.}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.lm", usually, a result of a call to summary.lm.}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}
}
\description{
\code{summary} method for class "\code{awesome_lm}"
}
\details{
`summary.awesome_lm` is based on \code{summary}.
}
\examples{

# Annette Dobson (1990) "An Introduction to Generalized Linear Models".
# Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)

# Example 1: Standard regresssion
lm1 <- lm(weight ~ group)
summary(lm1)

# Example 2: 90\% confidence intervals
lm2 <- lm(weight ~ group, level = .9)
summary(lm2)

}
\seealso{
\code{awesome_lm}.
}
