% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_describe.r
\name{tidy_describe}
\alias{tidy_describe}
\title{Convert descriptives to a tidy data frame}
\usage{
tidy_describe(descriptives)
}
\arguments{
\item{descriptives}{A descriptives data frame created with tidystats' \code{describe}.}
}
\description{
\code{tidy_describe} returns a tidy data frame of descriptive statistics created with \strong{tidystats}' \code{describe}.
}
\examples{
library(dplyr)

# Get descriptives
descriptives <- describe(sleep, extra)

# Create a tidy data frame of the descriptives
tidy_describe(descriptives)

# With a grouping variable:
sleep \%>\%
  group_by(group) \%>\%
  describe(extra) \%>\%
  tidy_describe()

}
