% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_descriptives.r
\name{tidy_descriptives}
\alias{tidy_descriptives}
\title{tidy_descriptives Function}
\usage{
tidy_descriptives(descriptives)
}
\description{
\code{tidy_descriptives} returns a tidy data frame of descriptive statistics.
}
\examples{
results <- descriptives(sleep, group)
tidy_descriptives(results)

}
