% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_frequencies.r
\name{tidy_frequencies}
\alias{tidy_frequencies}
\title{tidy_frequencies Function}
\usage{
tidy_frequencies(frequencies)
}
\description{
\code{tidy_frequencies} returns a tidy data frame of frequency statistics.
}
\examples{
results <- frequencies(sleep, group)
tidy_frequencies(results)

}
