% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{tidy_matrix}
\alias{tidy_matrix}
\title{Helper functions in tidystats}
\usage{
tidy_matrix(m)
}
\arguments{
\item{m}{A matrix.}
}
\description{
Functions used under the hood in the \code{tidystats} package.
}
\section{Functions}{
\itemize{
\item \code{tidy_matrix}: Function to convert matrix objects to a tidy data frame.
}}

