% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.afex_aov.r
\name{tidy_stats.afex_aov}
\alias{tidy_stats.afex_aov}
\title{Create a tidy stats data frame from an afex aov object}
\usage{
\method{tidy_stats}{afex_aov}(model)
}
\arguments{
\item{model}{Output of afex's \code{aov_ez}, \code{aov_car}, or \code{aov_4}}
}
\description{
\code{tidy_stats.aov} takes an afex aov object and converts the object to a
tidy stats data frame.
}
\examples{
# Load data
data(obk.long, package = "afex")

# Conduct an ANOVA
afex_aov <- aov_4(value ~ treatment * gender + (phase * hour | id),
  data = obk.long, observed = "gender")

# Tidy stats
tidy_stats(afex_aov)

}
