% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.aovlist.R
\name{tidy_stats.aovlist}
\alias{tidy_stats.aovlist}
\title{tidy_stats method for an aovlist object}
\usage{
\method{tidy_stats}{aovlist}(model)
}
\description{
Creates a tidystats data frame for an aovlist object.
}
