% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.confint.r
\name{tidy_stats.confint}
\alias{tidy_stats.confint}
\title{tidy_stats method for confint output.}
\usage{
\method{tidy_stats}{confint}(model, args = NULL)
}
\arguments{
\item{model}{The output of confint().}

\item{args}{Unused.}
}
\description{
Creates a tidystats data frame from the output of confint.
}
\details{
This method should not be called directly.
}
