% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.emmGrid.R
\name{tidy_stats.emmGrid}
\alias{tidy_stats.emmGrid}
\title{Create a tidy stats data frame from an emmGrid object}
\usage{
\method{tidy_stats}{emmGrid}(model)
}
\arguments{
\item{model}{Output of emmeans's \code{emmeans} without pairwise comparisons.}
}
\description{
\code{tidy_stats.emmGrid} takes an emmGrid object and converts the object to a
tidy stats data frame.
}
\examples{
# Load data
pigs <- emmeans::pigs

# Conduct a linear regression
pigs.lm1 <- lm(log(conc) ~ source + factor(percent), data = pigs)

# Tidy stats
tidy_stats(pigs.lm1)

}
