% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lm.r
\name{tidy_stats.lm}
\alias{tidy_stats.lm}
\title{Create a tidy stats json object from an lm object}
\usage{
\method{tidy_stats}{lm}(model)
}
\description{
Create a tidy stats json object from an lm object
}
