% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lme4.R
\name{tidy_stats.lmerMod}
\alias{tidy_stats.lmerMod}
\title{tidy_stats method for lmerMod objects}
\usage{
\method{tidy_stats}{lmerMod}(model)
}
\arguments{
\item{model}{An lmerMod object}
}
\description{
Creates a tidystats data frame for an lmerMod object.
}
\examples{
lme4_model <- lmer(extra ~ group + (1|ID), data = sleep)
tidy_stats.lmerMod(lme4_model)

}
