% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.merModLmerTest.R
\name{tidy_stats.merModLmerTest}
\alias{tidy_stats.merModLmerTest}
\title{tidy_stats method for merModLmerTest objects}
\usage{
\method{tidy_stats}{merModLmerTest}(model)
}
\arguments{
\item{model}{An merModLmerTest object}
}
\description{
Creates a tidystats data frame for an merModLmerTest object.
}
