% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.R
\name{tidy_stats}
\alias{tidy_stats}
\title{Create a tidy stats json object from a statistical output object}
\usage{
tidy_stats(model, args = NULL)
}
\description{
Create a tidy stats json object from a statistical output object
}
