% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.rma.R
\name{tidy_stats.rma}
\alias{tidy_stats.rma}
\title{Create a tidy stats data frame from an rma object from the metafor package}
\usage{
\method{tidy_stats}{rma}(model)
}
\arguments{
\item{model}{Output of \code{metafor::rma()}.}
}
\description{
\code{tidy_stats.rma} takes an rma object and converts the object to a tidy stats data frame.
}
