% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.summary.glm.r
\name{tidy_stats.summary.glm}
\alias{tidy_stats.summary.glm}
\title{tidy_stats method for summary.glm objects}
\usage{
\method{tidy_stats}{summary.glm}(model)
}
\arguments{
\item{model}{A summary.glm object}
}
\description{
Creates a tidystats data frame for a summary.glm object.
}
