% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.summary.manova.r
\name{tidy_stats.summary.manova}
\alias{tidy_stats.summary.manova}
\title{tidy_stats method for manova objects}
\usage{
\method{tidy_stats}{summary.manova}(model)
}
\arguments{
\item{model}{A manova object}
}
\description{
Creates a tidystats data frame for a manova object.
}
