% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.tidystats_descriptives.R
\name{tidy_stats.tidystats_descriptives}
\alias{tidy_stats.tidystats_descriptives}
\title{Convert a tidystats_descriptives data frame to a list}
\usage{
\method{tidy_stats}{tidystats_descriptives}(data)
}
\arguments{
\item{data}{A data frame created with \code{describe_data}.}
}
\description{
\code{tidy_stats.tidystats_descriptives} converts a data frame created with
\code{describe_data} to a list that can, in turn, be converted to a .json 
file.
}
\details{
When the data frame also contains grouping variables, please make
sure that the data frame is grouped by those variables using \strong{dplyr}'s 
\code{group_by}. This is automatically the case when the output of 
\code{describe_data} is used.
}
