% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats_confint.r
\name{tidy_stats_confint}
\alias{tidy_stats_confint}
\title{tidy_stats method for confint output.}
\usage{
tidy_stats_confint(confint_matrix, method = NULL)
}
\arguments{
\item{matrix}{confint matrix.}
}
\description{
Creates a tidystats data frame from the output of confint.
}
