% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_total.r
\name{tidy_total}
\alias{tidy_total}
\title{Convert count descriptives to a tidy data frame}
\usage{
tidy_total(descriptives)
}
\arguments{
\item{descriptives}{a descriptives data frame created with tidystats' \code{total}.}
}
\description{
\code{tidy_total} returns a tidy data frame of count statistics created with \strong{tidystats}' \code{total}.
}
\examples{
library(dplyr)

# Get descriptives
condition_n <- total(cox, condition)

# Create a tidy data frame of the descriptives
tidy_total(condition_n)

# With a grouping variable:
cox \%>\%
  group_by(sex) \%>\%
  describe(condition) \%>\%
  tidy_total()

}
