% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystats.aov.r
\name{tidystats.aov}
\alias{tidystats.aov}
\title{Tidystats method for an aov object}
\usage{
tidystats.aov(model, identifier, type = "other", description = NULL)
}
\description{
Creates a tidystats data.frame for an aov object.
}
