% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystats.lm.r
\name{tidystats.lm}
\alias{tidystats.lm}
\title{Tidystats method for an lm object}
\usage{
\method{tidystats}{lm}(model, identifier, type = "", description = NULL)
}
\description{
Creates a tidystats data.frame for an lm object.
}
\examples{
lm_model <- lm(extra ~ group, data = sleep)
tidystats.htest("model1", "hypothesis", lm_model, "Linear model to test the effect of group on extra.")

}
