% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.r
\name{total}
\alias{total}
\title{Calculate the total of observations}
\usage{
total(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables you want to total}
}
\description{
\code{total} returns the number of observations for categorical variables.
}
\details{
The data set can be grouped using \code{group_by} so that total of observations will be calculated for each group level.
}
\examples{
library(magrittr)
library(dplyr)

# 1 variable
total(cox, condition)

# 1 variable, 1 group
cox \%>\%
  group_by(sex) \%>\%
  describe(condition)

}
