% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{translate_statistic}
\alias{translate_statistic}
\title{Translate statistic column}
\usage{
translate_statistic(tibble, dictionary)
}
\arguments{
\item{tibble}{A tibble with a column called 'statistic'}

\item{dictionary}{A dictionary}
}
\value{
The provided tibble with the values in the column
    'statistic' replaced according to the dictionary
}
\description{
Very often statistical methods use their own esoteric names for
common (and less common) statistical methods. This method
translates those esoteric names to hopefully more sensible names
using a dictionary.
}
