% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_descriptives.R
\name{write_descriptives}
\alias{write_descriptives}
\title{Write a csv file containing descriptives}
\usage{
write_descriptives(descriptives, path)
}
\description{
\code{write_descriptives} creates a .csv file containing descriptives
}
\examples{
descriptives <- list()

descriptives <- sleep \%>\%
  describe(variable = "measure", group = "group") \%>\%
  add_descriptives(descriptives, identifier = "sleep_data")

write_descriptives(descriptives, "descriptives.csv")

}
