% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{write_stats}
\alias{write_stats}
\title{Write a tidy stats list to a file}
\usage{
write_stats(results, path)
}
\arguments{
\item{results}{A tidy stats list.}

\item{path}{Path or connection to write to.}
}
\description{
\code{write_stats} writes a tidy stats list to a .json file.
}
