% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_descriptives.R
\name{add_descriptives}
\alias{add_descriptives}
\title{add_descriptives Function}
\usage{
add_descriptives(descriptives, list, identifier = NULL, description = NULL)
}
\description{
\code{add_descriptives} is a generic function to add descriptive information to a tidy descriptives data frame.
}
\examples{
descriptives <- list()

# Use Student's sleep data
sleep \%>\%
  group_by(group) \%>\%
  summarize(
    mean = mean(extra),
    sd = sd(extra),
    n = n()
  ) -> sleep_descriptives

descriptives <- add_descriptives(descriptives, sleep_descriptives,
                                 identifier = "D1")

}
