% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_descriptives.r
\name{add_descriptives}
\alias{add_descriptives}
\title{Add descriptives to a tidy stats list}
\usage{
add_descriptives(descriptives, results, identifier = NULL,
  statistics = NULL, type = NULL, confirmatory = NULL, notes = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model. Automatically created if not provided.}

\item{statistics}{A vector of statistics to select from the output and add to the tidy stats list.}

\item{notes}{A character string to add additional information. Some statistical tests produce notes information, which will be overwritten if notes are provided.}

\item{data}{A data frame that contains descriptives, created with \code{describe} or \code{total}.}
}
\description{
\code{add_descriptives} adds descriptives to a tidy results list. Descriptives are structured differently than output from statistical models, so it is recommended to create separate lists.
}
\examples{

# Create an empty list to store the results in
results <- list()

# Example: Manual chi-squared test of independence
x_squared_data <- data.frame(
  statistic = c("X-squared", "df", "p"),
  value = c(5.4885, 6, 0.4828),
  method = "Chi-squared test of independence"
  )

# Add results to results
results <- add_stats(x_squared_data, results, identifier = "M1")

}
