% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.data.frame.r
\name{add_stats.data.frame}
\alias{add_stats.data.frame}
\title{add_stats data frame function}
\usage{
\method{add_stats}{data.frame}(output, results, identifier = NULL,
  statistics = NULL, type = NULL, confirmatory = NULL, notes = NULL)
}
\description{
\code{add_stats.data.frame} is a function to add a tidy data frame of results to a tidy stats list. tidystats does not support all possible statistical tests, so it may not be able to produce tidy output of a statistical model. The best solution for now is to tidy the output of a statistical method yourself, creating a tidy data frame, and then use \code{add_stats}, which will call this function, to add it to the tidy stats list.
}
