% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.default.r
\name{add_stats.default}
\alias{add_stats.default}
\title{add_stats default function}
\usage{
\method{add_stats}{default}(results, output, identifier = NULL,
  type = NULL, preregistered = NULL, notes = NULL, class = NULL,
  args = NULL)
}
\arguments{
\item{results}{A tidystats list.}

\item{output}{Output of a statistical test.}

\item{identifier}{A character string identifying the model. Automatically
created if not provided.}

\item{type}{A character string specifying the type of analysis: primary, 
secondary, or exploratory.}

\item{preregistered}{A boolean specifying whether the analysis was 
preregistered or not.}

\item{notes}{A character string specifying additional information.}
}
\description{
\code{add_stats.default} is the default add_stats() function. It takes the
output of a statistical test, organizes the output, and adds it to a 
tidystats list. While adding the output, additional information about the 
test can be supplied, including the type of test (primary, secondary, or 
exploratory), whether the test was preregistered, and additional notes.
}
\details{
Please note that not all statistical tests are currently supported. See 
'Details' below for a list of supported statistical tests.


Currently supported functions:
- t.test()
- cor.test()
- chisq.test()
- wilcox.test()
- fisher.test()
- oneway.test()
- aov()
- lm()
}
