% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.default.r
\name{add_stats.default}
\alias{add_stats.default}
\title{add_stats generic function}
\usage{
\method{add_stats}{default}(output, results, identifier = NULL,
  statistics = NULL, type = NULL, confirmatory = NULL, notes = NULL)
}
\arguments{
\item{output}{output of a statistical test.}

\item{results}{a tidy stats list.}

\item{identifier}{a character string identifying the model. Automatically created if not provided.}

\item{type}{a character string indicating the type of test. One of "hypothesis", "manipulation check", "contrast", or "other", can be abbreviated.}

\item{description}{a character string to add additional information.}
}
\description{
\code{add_stats} is a generic function to add a line of tidy stats to a tidy stats data.frame.
}
\examples{
results <- list()
model <- t.test(1:10, y = c(7:20))
results <- add_stats(model, results, identifier = "M1", type = "hypothesis")

}
