% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.R
\name{add_stats}
\alias{add_stats}
\title{add_stats Generic Function}
\usage{
add_stats(results, model, identifier = NULL, type = "other",
  description = NULL)
}
\arguments{
\item{results}{a tidy stats list.}

\item{model}{output of a statistical test.}

\item{identifier}{a character string identifying the model. Automatically created if not provided.}

\item{type}{a character string indicating the type of test. One of "hypothesis", "manipulation check", "contrast", or "other", can be abbreviated.}

\item{description}{a character string to add additional information.}
}
\description{
\code{add_stats} is a generic function to add a line of tidy stats to a tidy stats data.frame.
}
\examples{
results <- list()
model <- t.test(1:10, y = c(7:20))
results <- add_stats(results, model, "M1", "hypothesis")
}
