% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.R
\name{add_stats}
\alias{add_stats}
\title{add_stats Generic Function}
\usage{
add_stats(results, model, identifier = NULL, type = "other",
  description = NULL)
}
\description{
\code{add_stats} is a generic function to add a line of tidy stats to a tidy stats data.frame.
}
\examples{
results <- new_stats_data_frame()
ttest <- t.test(1:10, y = c(7:20))
results <- add_stats(results, ttest, "ttest_x_y", "Hypothesis", "A t-test of x and y")
}
