% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.r
\name{add_stats}
\alias{add_stats}
\title{Add statistical output to a tidy stats list}
\usage{
add_stats(output, results, identifier = NULL, statistics = NULL,
  type = NULL, confirmatory = NULL, notes = NULL)
}
\arguments{
\item{output}{output of a statistical test or a data frame. If a data frame is provided, it must already be in a tidy format.}

\item{results}{a tidy stats list.}

\item{identifier}{a character string identifying the model. Automatically created if not provided.}

\item{statistics}{a vector of statistics to select from the output and add to the tidy stats list.}

\item{type}{a character string indicating the type of test. One of "hypothesis", "manipulation check", "contrast", "descriptives", or "other". Can be abbreviated.}

\item{confirmatory}{a boolean to indicate whether the statistical test was confirmatory (TRUE) or exploratory (FALSE). Can be NA.}

\item{notes}{a character string to add additional information. Some statistical tests produce notes information, which will be overwritten if notes are provided.}
}
\description{
\code{add_stats} adds output to a results list. It can take either the output of a statistical test as input or a data frame. See Details for more information on adding data frames.
}
\details{
Some statistical functions produce unidentifiable output, which means \code{tidystats} cannot figure out how to tidy the data. To add these results, the output should be manually tidied or tidied using one of \code{tidystats}'s tidy functions. See \code{?tidy_stats} for an overview of available functions.
}
\examples{
# Create an empty list to store the results in
results <- list()

# Conduct statistical tests
x <- rnorm(10, mean = 0)
y <- rnorm(10, mean = 1)

model1 <- t.test(x, y)
model2 <- cor.test(x, y)
model3 <- lm(y ~ x)

# Add output to the results list
results <- add_stats(model1, results, identifier = "M1",
                     statistics = c("t", "df", "p"), type = "h",
                     confirmatory = TRUE, notes = "An example t-test")
results <- add_stats(model2, results, identifier = "M2")
results <- add_stats(model3, results, identifier = "M3")

# Get confidence intervals of model3
ci <- confint(model3)

# Produce a tidy data frame of the CIs and add it to the results list
results <- ci \%>\%
  tidy_stats_confint() \%>\%
  add_stats(results, identifier = "M3_CIs")

}
