% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.r
\name{add_stats}
\alias{add_stats}
\title{Add statistical output to a tidy stats list}
\usage{
add_stats(output, results, identifier = NULL, statistics = NULL,
  type = NULL, confirmatory = NULL, notes = NULL)
}
\description{
\code{add_stats} is a generic function that adds output to a results list. It can take either a statistical model as input or a data frame.
}
