% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{describe}
\alias{describe}
\title{describe function}
\usage{
describe(data, variable, group = NULL, na.rm = TRUE)
}
\description{
\code{describe} is a function to produce descriptives.
}
\examples{
describe(sleep, extra)
describe(sleep, extra, group = group)

}
