% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect (a) statistical model(s) added to your tidystats list}
\usage{
inspect(results, ...)
}
\arguments{
\item{results}{A tidystats list.}

\item{...}{Models to show in the viewer.}
}
\description{
\code{inspect} is a function to inspect one or more models that are part of a
tidystats list. The function will open a Shiny widget in the Viewer pane,
which will show the results of one or more models. This allows the user to
visually inspect the model output, as well as copy the results in APA format.
}
