% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_descriptives.r
\name{M}
\alias{M}
\title{Report descriptives}
\usage{
M(identifier, var = NULL, group = NULL,
  results = getOption("tidystats_list"))
}
\description{
Report descriptives
}
\examples{
# Read in a list of results
descriptives <- read_stats(system.file("descriptives.csv", package = "tidystats"))

# Generic example:
D(descriptives, identifier = "D1", var = "avoidance", statistic = "M")

# Shorter examples:
# Mean
M(descriptives, "D1", "avoidance")

# SD
SD(descriptives, "D1", "avoidance")

}
