% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats.R
\name{read_stats}
\alias{read_stats}
\title{Read a .json file that was produced with \code{write_stats}}
\usage{
read_stats(file)
}
\arguments{
\item{file}{A string; specifying the path to the tidystats data file.}
}
\description{
\code{read_stats()} can read in a .json file containing the statistical output
that was produced with \code{\link[=write_stats]{write_stats()}}. It returns a list containing the
results, with the identifier as the name for each list element.
}
\examples{
# A simple example, assuming there is a file called 'results.json'
\dontrun{
results <- read_stats("results.json")
}

# A working example
results <- read_stats(system.file("results.json", package = "tidystats"))

}
