% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats.R
\name{read_stats}
\alias{read_stats}
\title{Read a .csv file with all stats}
\usage{
read_stats(file)
}
\description{
\code{read_stats} returns a list containing all stats, with the identifier as the name for each list element.
}
\examples{
results <- read_stats("results.csv")

}
