% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_anova.R
\name{report_anova}
\alias{report_anova}
\title{Report method for ANOVA models}
\usage{
report_anova(results, identifier, term = NULL, term_nr = NULL,
  statistic = NULL)
}
\description{
Function to report ANOVA output in APA style.
}
