% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_anova.R
\name{report_anova}
\alias{report_anova}
\title{Report method for ANOVA models}
\usage{
report_anova(results, identifier, group = NULL, term = NULL,
  term_nr = NULL, statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{term}{A character string indicating the term you want to report.}

\item{term_nr}{A number indicating the term you want to report.}

\item{statistic}{A character string identifying the exact statistic you want
to report.}
}
\description{
Function to report ANOVA output in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "ANOVA", term = "N")
report(results, identifier = "ANOVA", term_nr = "2")
report(results, identifier = "ANOVA", term = "N", statistic = "p")

}
