% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_correlation.R
\name{report_correlation}
\alias{report_correlation}
\title{Report function for correlations}
\usage{
report_correlation(results, identifier, term, term_nr)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{term}{A character string indicating the term you want to report.}

\item{term_nr}{A number indicating the term you want to report.}
}
\description{
Function to report a correlation in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "correlation_pearson")
report(results, identifier = "correlation_spearman")
report(results, identifier = "correlation_kendall")

}
