% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_correlation.R
\name{report_correlation}
\alias{report_correlation}
\title{Report method for correlations.}
\usage{
report_correlation(results, identifier, statistic = NULL)
}
\description{
Function to report a correlation in APA style.
}
\examples{
results <- new_stats_data_frame()

x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
y <- c( 2.6,  3.1,  2.5,  5.0,  3.6,  4.0,  5.2,  2.8,  3.8)

cor_test <- cor.test(x, y)

results <- add_stats(results, cor_test, "cor_test1")
report_correlation(results, "cor_test1")

}
