% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_correlation.R
\name{report_correlation}
\alias{report_correlation}
\title{Report method for correlations}
\usage{
report_correlation(results, identifier, statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{statistic}{A character string of a statistic you want to extract from a model.}
}
\description{
Function to report a correlation in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(identifier = "correlation", results = results)
report(identifier = "correlation", statistic = "p", results = results)

}
