% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_fisher.R
\name{report_fisher}
\alias{report_fisher}
\title{Report function for Fisher's Exact Tests for Count Data}
\usage{
report_fisher(results, identifier, statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{statistic}{A character string identifying the exact statistic you want
to report.}
}
\description{
Function to report Fisher's Exact Tests for Count Data in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "fisher_test")
report(results, identifier = "fisher_test", statistic = "OR")

}
