% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_lm.r
\name{report_lm}
\alias{report_lm}
\title{Report method for linear regression models}
\usage{
report_lm(results, identifier, group = NULL, term = NULL, term_nr = NULL,
  statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{group}{A character string indicating the group containing the statistic
or statistics you want to report.}

\item{term}{A character string indicating the term you want to report.}

\item{term_nr}{A number indicating the term you want to report.}

\item{statistic}{A character string identifying the exact statistic you want
to report.}
}
\description{
Function to report a regression in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Set the default results list
options(tidystats_list = results)

# Example: regression term
report("regression", term = "groupTrt")
report("regression", term_nr = 2)
report("regression", term = "groupTrt", statistic = "p")

}
