% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_lm.r
\name{report_lm}
\alias{report_lm}
\title{Report method for linear regression models}
\usage{
report_lm(results, identifier, term = NULL, term_nr = NULL,
  statistic = NULL)
}
\description{
Function to report a regression in APA style.
}
