% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p_value.R
\name{report_p_value}
\alias{report_p_value}
\title{Report p-value function}
\usage{
report_p_value(p_value)
}
\arguments{
\item{p_value}{A p-value.}
}
\description{
A helper function to report p values in APA style.
}
\examples{
report_p_value(0.532)
report_p_value(0.0432)
report_p_value(0.0000121)

}
