% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_rma.R
\name{report_rma}
\alias{report_rma}
\title{Report method for metafor's rma models}
\usage{
report_rma(identifier, group = NULL, term = NULL, term_nr = NULL,
  results = getOption("tidystats_list"))
}
\arguments{
\item{identifier}{A character string identifying the model.}

\item{group}{A character string identifying the group you want to report the
statistics of.}

\item{term}{A character string indicating which term you want to report the
statistics of.}

\item{term_nr}{A number indicating which term you want to report the the
statistics of.}

\item{results}{A tidystats list.}
}
\description{
Function to report a meta-analysis in APA style.
}
