% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t_test.R
\name{report_t_test}
\alias{report_t_test}
\title{Report function for t-tests}
\usage{
report_t_test(results, identifier)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}
}
\description{
Function to report a t-test in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "t_test_one_sample")
report(results, identifier = "t_test_two_sample")
report(results, identifier = "t_test_welch")
report(results, identifier = "t_test_paired")

}
