% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_table_lm.R
\name{report_table_lm}
\alias{report_table_lm}
\title{Report table method for linear regression models}
\usage{
report_table_lm(identifier, terms = NULL, term_nrs = NULL,
  statistics = NULL, include_model = TRUE, term_labels = NULL,
  results = getOption("tidystats_list"))
}
\arguments{
\item{identifier}{A character string identifying the model.}

\item{terms}{A character vector indicating which terms you want to report in the table.}

\item{term_nrs}{A numeric vector indicating which terms you want to report in the table.}

\item{statistics}{A character vector of statistics you want to display in the table.}

\item{include_model}{Include or exclude model statistics (e.g., R squared).}

\item{term_labels}{A character vector to change the labels for the terms.}

\item{results}{A tidystats list.}
}
\description{
Function create a table in order to display results of a regression.
}
