% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_wilcoxon.R
\name{report_wilcoxon}
\alias{report_wilcoxon}
\title{Report function for Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
report_wilcoxon(results, identifier, statistic = NULL)
}
\arguments{
\item{results}{A tidy stats list.}

\item{identifier}{A character string identifying the model.}

\item{statistic}{A character string identifying the exact statistic you want
to report.}
}
\description{
Function to report Wilcoxon Rank Sum and Signed Rank Tests in APA style.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Report results
report(results, identifier = "wilcoxon_signed_rank")
report(results, identifier = "wilcoxon_signed_rank", statistic = "Z")

}
