% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{stats}
\alias{stats}
\title{Create a custom statistics object}
\usage{
stats(method, statistics)
}
\arguments{
\item{method}{A string; specifying the path to type of method used to obtain
the statistics.}

\item{statistics}{A vector of statistics created with \code{stat()}.}
}
\description{
\code{stats} can be used to create your own statistics object to add to a
statistics list via \code{add_stats()}.
}
\examples{
# Example of a calculating a custom statistic and adding it to a tidystats
# list
lm1 <- lm(Fertility ~ ., data = swiss)
lm2 <- update(lm1, . ~ . - Examination)

BF10 <- 1 / exp((BIC(lm2) - BIC(lm1)) / 2)

stats <- stats(
  method = "BF BIC method",
  statistics = c(
    stat(name = "BF", value = BF10, subscript = "10"),
    stat(name = "BF", value = 1 / BF10, subscript = "01")
  )
)

statistics <- list()

statistics <- add_stats(
  list = statistics,
  output = stats,
  notes = "Wagenmakers (2007) method for calculating Bayes factors"
)

}
