% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_list_to_df.r
\name{stats_list_to_df}
\alias{stats_list_to_df}
\title{Convert a tidy stats list to a data frame}
\usage{
stats_list_to_df(results)
}
\arguments{
\item{results}{A tidy stats list.}
}
\description{
\code{stats_list_to_df} converts a tidy stats list to a data frame.
}
\examples{
# Read in a list of results
results <- read_stats(system.file("results.csv", package = "tidystats"))

# Convert list to a data frame
stats_list_to_df(results)

}
