% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.afex_aov.r
\name{tidy_stats.afex_aov}
\alias{tidy_stats.afex_aov}
\title{Create a tidy stats data frame from an afex aov object}
\usage{
\method{tidy_stats}{afex_aov}(model, args = NULL)
}
\arguments{
\item{model}{Output of afex's \code{aov_ez}, \code{aov_car}, or \code{aov_4}}

\item{args}{Unused}
}
\description{
\code{tidy_stats.aov} takes an afex aov object and converts the object to a
tidy stats data frame.
}
\examples{
# Check if afex package is available
if(!requireNamespace("afex", quietly = TRUE)) {

  message(paste0("Package 'afex' is needed for this example to work. ",
              "Please install it."), .call = FALSE) 
} else {

  # Load data
  data(obk.long, package = "afex")

  # Conduct an ANOVA
  afex_aov <- afex::aov_4(value ~ treatment * gender + (phase * hour | id),
                          data = obk.long, observed = "gender")

  # Tidy stats
  tidy_stats(afex_aov)
}

}
