% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.anova.r
\name{tidy_stats.anova}
\alias{tidy_stats.anova}
\title{tidy_stats method for anova objects}
\usage{
\method{tidy_stats}{anova}(model)
}
\arguments{
\item{model}{An anova object}
}
\description{
Creates a tidystats data frame for an anova object.
}
\examples{
lm1 <- lm(extra ~ 1, data = sleep)
tidy_stats.anova(lm1)

lm2 <- lm(extra ~ group, data = sleep)
tidy_stats.anova(lm1, lm2)

}
