% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.anova.r
\name{tidy_stats.anova}
\alias{tidy_stats.anova}
\title{Create a tidy stats data frame from an anova object}
\usage{
\method{tidy_stats}{anova}(model)
}
\arguments{
\item{model}{Output of \code{anova}.}
}
\description{
\code{tidy_stats.anova} takes an anova object and converts the object to a tidy stats data frame.
}
