% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.aov.r
\name{tidy_stats.aov}
\alias{tidy_stats.aov}
\title{Tidystats method for an aov object}
\usage{
\method{tidy_stats}{aov}(model, identifier, type = "other",
  description = NULL)
}
\description{
Creates a tidystats data.frame for an aov object.
}
