% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.aov.r
\name{tidy_stats.aov}
\alias{tidy_stats.aov}
\title{Create a tidy stats data frame from an aov object}
\usage{
\method{tidy_stats}{aov}(model, args = NULL)
}
\arguments{
\item{model}{Output of \code{aov} without within-subject factors.}

\item{args}{Unused.}
}
\description{
\code{tidy_stats.aov} takes an aov object and converts the object to a tidy stats data frame.
}
\examples{
# Conduct an ANOVA
model_aov <- aov(yield ~ block + N * P * K, npk)

# Tidy stats
tidy_stats(model_aov)

}
