% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.htest.R
\name{tidy_stats.htest}
\alias{tidy_stats.htest}
\title{Create a tidy stats data frame from an htest object}
\usage{
\method{tidy_stats}{htest}(model)
}
\arguments{
\item{model}{An htest object}
}
\description{
\code{tidy_stats.htest} takes an htest object and converts the object to a tidy stats data frame.
}
\examples{
tidy_stats.htest(t.test(1:10, y = c(7:20)))

}
