% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lm.r
\name{tidy_stats.lm}
\alias{tidy_stats.lm}
\title{tidy_stats method for an lm object}
\usage{
\method{tidy_stats}{lm}(model, identifier, type = "other",
  description = NULL)
}
\description{
Creates a tidy_stats data frame for an lm object.
}
