% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lm.r
\name{tidy_stats.lm}
\alias{tidy_stats.lm}
\title{Create a tidy stats data frame from an lm object}
\usage{
\method{tidy_stats}{lm}(model)
}
\arguments{
\item{model}{An lm object}
}
\description{
\code{tidy_stats.lm} takes an lm object and converts the object to a tidy stats data frame.
}
