% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lm.r
\name{tidy_stats.lm}
\alias{tidy_stats.lm}
\title{tidy_stats method for lm objects}
\usage{
\method{tidy_stats}{lm}(model)
}
\arguments{
\item{model}{An lm object}
}
\description{
Creates a tidystats data frame for an lm object.
}
\examples{
lm_model <- lm(extra ~ group, data = sleep)
tidy_stats.lm(lm_model)

}
