% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.lmerModLmerTest.r
\name{tidy_stats.lmerModLmerTest}
\alias{tidy_stats.lmerModLmerTest}
\title{Create a tidy stats data frame from an lmerModLmerTest object}
\usage{
\method{tidy_stats}{lmerModLmerTest}(model)
}
\arguments{
\item{model}{Output of lmerTest's \code{lmer()}.}
}
\description{
\code{tidy_stats.lmerModLmerTest} takes an lmerModLmerTest object and converts the object to a tidy stats data frame.
}
\examples{
# Check if lme4 and lmerTest packages are available
if(!requireNamespace("lme4", quietly = TRUE) &
   !requireNamespace("lmerTest", quietly = TRUE)) {

  message(paste0("Packages 'lme4' and 'lme4Test are needed for this ",
                 "example to work. Please install them."), .call = FALSE)
} else {

  # Conduct a linear mixed model
  model_lmerTest <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)

  # Tidy stats
  tidy_stats(model_lmerTest)
}

}
