% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.merModLmerTest.R
\name{tidy_stats.merModLmerTest}
\alias{tidy_stats.merModLmerTest}
\title{tidy_stats method for merModLmerTest objects}
\usage{
\method{tidy_stats}{merModLmerTest}(model)
}
\arguments{
\item{model}{An merModLmerTest object}
}
\description{
Creates a tidystats data frame for an merModLmerTest object.
}
\examples{
library(lme4)
library(lmerTest)

lme4_model <- lmer(extra ~ group + (1|ID), data = sleep)
tidy_stats.lmerMod(lme4_model)

}
