% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.mixed.R
\name{tidy_stats.mixed}
\alias{tidy_stats.mixed}
\title{Create a tidy stats data frame from an afex mixed object}
\usage{
\method{tidy_stats}{mixed}(model, args = NULL)
}
\arguments{
\item{model}{Output of afex's \code{mixed}.}
}
\description{
\code{tidy_stats.mixed} takes an afex mixed object and converts the object to
a tidy stats data frame.
}
\examples{
# Load data
data("sk2011.2")
sk2_aff <- droplevels(sk2011.2[sk2011.2$what == "affirmation",])

# Perform the analysis
sk_m1 <- mixed(response ~ instruction * inference * type +
  (inference * type | id), sk2_aff)

# Tidy stats
tidy_stats(sk_m1) # Default ANOVA output
tidy_stats(sk_m1, args = list(summary = "lmer")) # lmer summary

}
