% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.psych.R
\name{tidy_stats.psych}
\alias{tidy_stats.psych}
\title{tidy_stats method for psych's alpha objects}
\usage{
\method{tidy_stats}{psych}(model)
}
\arguments{
\item{model}{An psych alpha object}
}
\description{
Creates a tidystats data frame for a psych's alpha object.
}
\examples{
library(psych)
results <- list()

bfi \%>\%
  select(A1, A2, A3, A4, A5) \%>\%
  alpha(check.keys = TRUE) \%>\%
  add_stats(results, identifier = 'alpha_agreeableness')

}
