% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.psych.R
\name{tidy_stats.psych}
\alias{tidy_stats.psych}
\title{tidy_stats method for psych's alpha objects}
\usage{
\method{tidy_stats}{psych}(model)
}
\arguments{
\item{model}{An alpha object}
}
\description{
Creates a tidystats data frame for a psych's alpha object.
}
