% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats.BFBayesFactor.r,
%   R/tidy_stats.Hmisc.r, R/tidy_stats.R, R/tidy_stats.afex.r,
%   R/tidy_stats.brmsfit.r, R/tidy_stats.effectsize.r, R/tidy_stats.effsize.r,
%   R/tidy_stats.emmeans.R, R/tidy_stats.irr.r, R/tidy_stats.lavaan.r,
%   R/tidy_stats.lmerMod.r, R/tidy_stats.nlme.r, R/tidy_stats.psych.r,
%   R/tidy_stats.tidystats.r
\name{tidy_stats.BFBayesFactor}
\alias{tidy_stats.BFBayesFactor}
\alias{tidy_stats.rcorr}
\alias{tidy_stats}
\alias{tidy_stats.htest}
\alias{tidy_stats.pairwise.htest}
\alias{tidy_stats.lm}
\alias{tidy_stats.glm}
\alias{tidy_stats.anova}
\alias{tidy_stats.aov}
\alias{tidy_stats.aovlist}
\alias{tidy_stats.confint}
\alias{tidy_stats.afex_aov}
\alias{tidy_stats.mixed}
\alias{tidy_stats.brmsfit}
\alias{tidy_stats.effectsize_difference}
\alias{tidy_stats.effsize}
\alias{tidy_stats.emmGrid}
\alias{tidy_stats.summary_emm}
\alias{tidy_stats.emm_list}
\alias{tidy_stats.icclist}
\alias{tidy_stats.lavaan}
\alias{tidy_stats.lmerMod}
\alias{tidy_stats.lmerModLmerTest}
\alias{tidy_stats.lme}
\alias{tidy_stats.nlme}
\alias{tidy_stats.anova.lme}
\alias{tidy_stats.gls}
\alias{tidy_stats.psych}
\alias{tidy_stats.tidystats}
\alias{tidy_stats.tidystats_descriptives}
\alias{tidy_stats.tidystats_counts}
\title{Tidy the output of a statistics object}
\usage{
\method{tidy_stats}{BFBayesFactor}(x, args = NULL)

\method{tidy_stats}{rcorr}(x, args = NULL)

tidy_stats(x, args = NULL)

\method{tidy_stats}{htest}(x, args = NULL)

\method{tidy_stats}{pairwise.htest}(x, args = NULL)

\method{tidy_stats}{lm}(x, args = NULL)

\method{tidy_stats}{glm}(x, args = NULL)

\method{tidy_stats}{anova}(x, args = NULL)

\method{tidy_stats}{aov}(x, args = NULL)

\method{tidy_stats}{aovlist}(x, args = NULL)

\method{tidy_stats}{confint}(x, args = NULL)

\method{tidy_stats}{afex_aov}(x, args = NULL)

\method{tidy_stats}{mixed}(x, args = NULL)

\method{tidy_stats}{brmsfit}(x, args = NULL)

\method{tidy_stats}{effectsize_difference}(x, args = NULL)

\method{tidy_stats}{effsize}(x, args = NULL)

\method{tidy_stats}{emmGrid}(x, args = NULL)

\method{tidy_stats}{summary_emm}(x, args = NULL)

\method{tidy_stats}{emm_list}(x, args = NULL)

\method{tidy_stats}{icclist}(x, args = NULL)

\method{tidy_stats}{lavaan}(x, args = NULL)

\method{tidy_stats}{lmerMod}(x, args = NULL)

\method{tidy_stats}{lmerModLmerTest}(x, args = NULL)

\method{tidy_stats}{lme}(x, args = NULL)

\method{tidy_stats}{nlme}(x, args = NULL)

\method{tidy_stats}{anova.lme}(x, args = NULL)

\method{tidy_stats}{gls}(x, args = NULL)

\method{tidy_stats}{psych}(x, args = NULL)

\method{tidy_stats}{tidystats}(x, args = NULL)

\method{tidy_stats}{tidystats_descriptives}(x, args = NULL)

\method{tidy_stats}{tidystats_counts}(x, args = NULL)
}
\arguments{
\item{x}{The output of a statistical test.}
}
\description{
\code{tidy_stats} is used to convert the output of a statistical object to a
list of organized statistics. The \code{tidy_stats} function is automatically
run when \code{add_stats} is used, so there is generally no need to use this
function explicitly. It can be used, however, to peek at how the output of a
specific analysis will be organized.
}
\section{Methods (by class)}{
\itemize{
\item \code{tidy_stats(BFBayesFactor)}: tidy_stats method for class 'BayesFactor'

\item \code{tidy_stats(rcorr)}: tidy_stats method for class 'rcorr'

\item \code{tidy_stats(htest)}: tidy_stats method for class 'htest'

\item \code{tidy_stats(pairwise.htest)}: tidy_stats method for class 'pairwise.htest'

\item \code{tidy_stats(lm)}: tidy_stats method for class 'lm'

\item \code{tidy_stats(glm)}: tidy_stats method for class 'glm'

\item \code{tidy_stats(anova)}: tidy_stats method for class 'anova'

\item \code{tidy_stats(aov)}: tidy_stats method for class 'aov'

\item \code{tidy_stats(aovlist)}: tidy_stats method for class 'aovlist'

\item \code{tidy_stats(confint)}: tidy_stats method for class 'confint'

\item \code{tidy_stats(afex_aov)}: tidy_stats method for class 'afex_aov'

\item \code{tidy_stats(mixed)}: tidy_stats method for class 'afex_aov'

\item \code{tidy_stats(brmsfit)}: tidy_stats method for class 'brmsfit'

\item \code{tidy_stats(effectsize_difference)}: tidy_stats method for class 'effectsize_difference'

\item \code{tidy_stats(effsize)}: tidy_stats method for class 'effsize'

\item \code{tidy_stats(emmGrid)}: tidy_stats method for class 'emmGrid'

\item \code{tidy_stats(summary_emm)}: tidy_stats method for class 'summary_emm'

\item \code{tidy_stats(emm_list)}: tidy_stats method for class 'emm_list'

\item \code{tidy_stats(icclist)}: tidy_stats method for class 'icclist'

\item \code{tidy_stats(lavaan)}: tidy_stats method for class 'lavaan'

\item \code{tidy_stats(lmerMod)}: tidy_stats method for class 'lmerMod'

\item \code{tidy_stats(lmerModLmerTest)}: tidy_stats method for class 'lmerModLmerTest'

\item \code{tidy_stats(lme)}: tidy_stats method for class 'lme'

\item \code{tidy_stats(nlme)}: tidy_stats method for class 'nlme'

\item \code{tidy_stats(anova.lme)}: tidy_stats method for class 'anova.lme'

\item \code{tidy_stats(gls)}: tidy_stats method for class 'gls'

\item \code{tidy_stats(psych)}: tidy_stats method for class 'psych'

\item \code{tidy_stats(tidystats)}: tidy_stats method for class 'tidystats'

\item \code{tidy_stats(tidystats_descriptives)}: tidy_stats method for class 'tidystats_descriptives'

\item \code{tidy_stats(tidystats_counts)}: tidy_stats method for class 'tidystats_counts'

}}
\keyword{internal}
