% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats_to_data_frame.R
\name{tidy_stats_to_data_frame}
\alias{tidy_stats_to_data_frame}
\title{Convert a tidy stats list to a data frame}
\usage{
tidy_stats_to_data_frame(x)
}
\arguments{
\item{results}{A tidy stats list.}
}
\description{
\code{tidy_stats_to_data_frame} converts a tidy stats list to a data frame.
}
