% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystats.htest.R
\name{tidystats.htest}
\alias{tidystats.htest}
\title{Tidystats method for an htest object}
\usage{
\method{tidystats}{htest}(model, identifier, type = "other",
  description = NULL)
}
\description{
Creates a tidystats data.frame for an htest object.
}
\examples{
ttest <- t.test(1:10, y = c(7:20))
tidystats.htest("ttest_x_y", "Hypothesis", ttest, "A t-test for x and y")

}
