% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystats.R
\name{tidystats}
\alias{tidystats}
\title{Tidystats generic function}
\usage{
tidystats(model, identifier, type = "", description = NULL)
}
\description{
\code{tidystats} is a generic function that returns a line of tidy stats.
}
\examples{
tidystats.t.test()
}
