% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.r
\name{total}
\alias{total}
\title{Calculate the total of observations}
\usage{
total(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{the variables you want to total.}
}
\description{
\code{total} returns the number of observations for categorical variables.
}
\details{
The data set can be grouped using \strong{dplyr}'s \code{group_by} so that the total of observations will be calculated for each group level.
}
\examples{
library(tidyverse)

# 1 variable
total(cox, sex)

# 1 variable, 1 group
cox \%>\%
  group_by(condition) \%>\%
  describe(sex)

}
