% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{write_stats}
\alias{write_stats}
\title{Write a csv file with all stats}
\usage{
write_stats(results, path)
}
\description{
\code{write_stats} creates a .csv file with all stats
}
\examples{
results <- list()
model <- t.test(1:10, y = 7:20)
results <- add_stats(results, model)
write_stats(results, "results.csv")

}
