% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{write_stats}
\alias{write_stats}
\title{Write a .csv file with all stats}
\usage{
write_stats(results, file)
}
\description{
\code{write_stats} creates a \code{.csv} file with all stats
}
\examples{
results <- new_stats_list()
model <- t.test(1:10, y = c(7:20))
results <- add_stats(results, model, "model1", "hypothesis")
write_stats(results)
}
